some cleanup and reorganization
