+++ src/apps/chem/CMakeLists.txt
@@ -13,7 +13,7 @@ set(MADCHEM_HEADERS
     pcm.h SCFProtocol.h AC.h pointgroupoperator.h pointgroupsymmetry.h MolecularOrbitals.h 
 	znemo.h zcis.h  diamagneticpotentialfactor.h oep.h BSHApply.h
     PNO.h PNOF12Potentials.h PNOGuessFunctions.h PNOParameters.h PNOStructures.h PNOTensors.h
-    basis.h polynomial.h gaussian.h ESInterface.h NWChem.h orbital_partitioner.h)
+    basis.h polynomial.h gaussian.h ESInterface.h NWChem.h )
 set(MADCHEM_SOURCES
     correlationfactor.cc molecule.cc molecularbasis.cc vibanal.cc
     corepotential.cc atomutil.cc lda.cc
@@ -23,7 +23,7 @@ set(MADCHEM_SOURCES
     znemo.cc zcis.cc diamagneticpotentialfactor.cc oep.cc
     PNO.cpp PNOF12Potentials.cpp PNOGuessFunctions.cpp PNOParameters.cpp PNOStructures.cpp
     QCCalculationParametersBase.cc MolecularOrbitals.cc
-    polynomial.cc gaussian.cc NWChem.cc orbital_partitioner.cc)
+    polynomial.cc gaussian.cc NWChem.cc )
 if(LIBXC_FOUND)
   list(APPEND MADCHEM_SOURCES xcfunctional_libxc.cc)
 else()
