+++ src/madness/mra/test_vectormacrotask.cc
@@ -39,6 +39,7 @@
 #include <madness/mra/macrotaskq.h>
 #include <madness/world/world.h>
 #include <madness/world/timing_utilities.h>
+#include <madness/mra/macrotaskpartitioner.h>
 
 
 using namespace madness;
@@ -68,19 +69,8 @@ struct gaussian {
         double r2 = inner(r,r);
         return exp(-a * r2);//*abs(sin(abs(2.0*x))) *cos(y);
     }
-
 };
 
-
-static std::atomic<int> atomic_idx;
-
-
-template<typename ... Ts>
-constexpr auto decay_types(std::tuple<Ts...> const &)
--> std::tuple<std::remove_cv_t<std::remove_reference_t<Ts>>...>;
-
-template<typename T>
-using decay_tuple = decltype(decay_types(std::declval<T>()));
 
 template<typename Q>
 struct is_vector : std::false_type {
@@ -89,128 +79,17 @@ template<typename Q>
 struct is_vector<std::vector<Q>> : std::true_type {
 };
 
-template<typename>
-struct is_madness_function_vector : std::false_type {
-};
-
-template<typename T, std::size_t NDIM>
-struct is_madness_function_vector<std::vector<typename madness::Function<T, NDIM>>> : std::true_type {
-};
-
-
-template<size_t I = 0, typename tupleT>
-constexpr std::size_t get_index_of_first_vector_argument() {
-
-    typedef decay_tuple<tupleT> argtupleT;   // removes const, &, etc
-
-    if constexpr(I >= std::tuple_size_v<tupleT>) {
-        // Last case, if nothing is left to iterate, then exit the function
-        MADNESS_EXCEPTION("there is no madness function vector argument in the list, cannot partition the tasks",1);
-        return I;
-    } else {
-        // Print the tuple and go to next element
-        using typeT = typename std::tuple_element<I,argtupleT>::type;       // use decay types for determining a vector
-        if constexpr (is_madness_function_vector<typeT>::value) {
-            return I;
-        } else {
-            // Going for next element.
-            return get_index_of_first_vector_argument<I + 1,tupleT>();
-        }
-    }
-}
-
-
-class Batch {
-    friend class OrbitalPartitioner;
-public:
-    std::size_t begin, end; ///< first and first past last index [begin,end)
-
-    Batch(const std::size_t& begin, const std::size_t& end) : begin(begin), end(end) {}
-
-    /// select the relevant vector elements from the argument tuple
-    template<typename tupleT>
-    tupleT operator()(const tupleT &arg) const {
-
-        constexpr std::size_t index=get_index_of_first_vector_argument<0,tupleT>();
-        auto v=std::get<index>(arg);
-        decltype(v) v_batch;
-        std::copy(v.begin()+begin,v.begin()+end,std::back_inserter(v_batch));
-        tupleT batched_arg=arg;
-        std::get<index>(batched_arg)=v_batch;
-        return batched_arg;
-    }
-
-    friend std::ostream& operator<<(std::ostream& os, const Batch& batch) {
-        std::stringstream ss;
-        ss << "[" << std::setw(3) << batch.begin << ", " << batch.end << ")";
-        os << ss.str();
-        return os;
-    }
-
-};
-
-class OrbitalPartitioner {
-    friend class Batch;
-public:
-    typedef std::list<Batch> partitionT;
-    std::size_t min_batch_size;
-    std::size_t max_batch_size=10;
-    std::size_t nsubworld;
-    std::string policy="guided";
-
-    OrbitalPartitioner(long nsubworld, long min_batch_size=5, std::string policy="guided")
-            : min_batch_size(min_batch_size), nsubworld(nsubworld), policy(policy) {}
-
-    template<typename tupleT>
-    partitionT partition_tasks(const tupleT &argtuple) const {
-
-        auto v=get_first_vector_argument(argtuple);
-        partitionT result;
-
-        if (policy == "guided") {
-            std::size_t begin=0;
-            std::size_t end=0;
-            while (end<v.size()) {
-                end+=std::min(max_batch_size,std::max(min_batch_size,((v.size()-end)/nsubworld)));
-                end=std::min(end,v.size());
-                result.push_back(Batch(begin,end));
-                begin=end;
-            }
-        } else {
-            std::string msg="unknown partitioning policy: "+policy;
-            MADNESS_EXCEPTION(msg.c_str(),1);
-        }
-        return result;
-    }
-
-
-    /// find the first madness function vector argument in tuple and return it
-    template<size_t I = 0, typename... Ts>
-    static constexpr auto get_first_vector_argument(std::tuple<Ts...> argtuple) {
-
-        typedef decay_tuple<typename std::tuple<Ts...>> argtupleT;   // removes const, &, etc
-
-        if constexpr(I == sizeof...(Ts)) {
-            // Last case, if nothing is left to iterate, then exit the function
-            MADNESS_EXCEPTION("there is no madness function vector argument in the list, cannot partition the tasks",1);
-            return double(0.);
-        } else {
-            // Print the tuple and go to next element
-            using typeT = typename std::tuple_element<I,argtupleT>::type;       // use decay types for determining a vector
-            if constexpr (is_madness_function_vector<typeT>::value) {
-                return std::get<I>(argtuple);
-            } else {
-                // Going for next element.
-                return get_first_vector_argument<I + 1>(argtuple);
-            }
-        }
-    }
-};
 
 
 template<typename taskT>
 class MacroTask_2G {
-    using partitionT = OrbitalPartitioner::partitionT;
+    using partitionT = MacroTaskPartitioner::partitionT;
+
+//    template<typename ... Ts>
+//    constexpr auto decay_types(std::tuple<Ts...> const &)
+//    -> std::tuple<std::remove_cv_t<std::remove_reference_t<Ts>>...>;
+//
+//    template<typename T> using decay_tuple = decltype(decay_types(std::declval<T>()));
     typedef typename taskT::resultT resultT;
     typedef typename taskT::argtupleT argtupleT;
     typedef Cloud::recordlistT recordlistT;
@@ -231,7 +110,7 @@ public:
         static_assert(std::is_same<decltype(argtuple), argtupleT>::value, "type or number of arguments incorrect");
 
         // partition the argument vector into batches
-        OrbitalPartitioner op(taskq_ptr->get_nsubworld());
+        MacroTaskPartitioner op(taskq_ptr->get_nsubworld());
         partitionT partition = op.partition_tasks(argtuple);
 
         // store input and output: output being a pointer to a universe function (vector)
@@ -361,18 +240,12 @@ public:
         return result;
     }
 
-    std::pair<long, hashT> partition(const argtupleT &argtuple) {
-        return std::make_pair<long, hashT>(0, 0);
-    }
-
     resultT operator()(const real_function_3d &f1, const double &arg2,
                        const std::vector<real_function_3d> &f2) const {
         World& world=f1.world();
         real_convolution_3d op=CoulombOperator(world,1.e-4,1.e-5);
         return arg2 * f1 * apply(world,op,f2);
     }
-
-    real_function_3d density;
 };
 
 
@@ -391,10 +264,6 @@ public:
         return result;
     }
 
-    std::pair<long, hashT> partition(const argtupleT &argtuple) {
-        return std::make_pair<long, hashT>(0, 0);
-    }
-
     resultT operator()(const real_function_3d &f1, const double &arg2,
                        const std::vector<real_function_3d> &f2) const {
         World &world = f1.world();
@@ -403,16 +272,6 @@ public:
         return result;
     }
 };
-
-int test_orbital_partitioner(World &world) {
-    std::tuple<double, int, std::vector<real_function_4d>, std::vector<real_function_3d>> tuple;
-    auto v=OrbitalPartitioner::get_first_vector_argument(tuple);
-    constexpr std::size_t index=get_index_of_first_vector_argument<0,decltype(tuple)>();
-    print("test_orbital_partitioner, index ",index);
-
-    return 1;
-}
-
 
 int main(int argc, char **argv) {
     madness::World &universe = madness::initialize(argc, argv);
@@ -426,7 +285,6 @@ int main(int argc, char **argv) {
     int nworld = universe.size();
     if (universe.rank() == 0) print("creating nworld", nworld, universe.id());
 
-    // TODO: compute batches
     // TODO: serialize member variables of tasks
     // TODO: pretty-print cloud content/input/output records
     {
@@ -518,7 +376,6 @@ int main(int argc, char **argv) {
             else print("\n --> all tests \033[31m", "failed \033[0m \n");
         }
     }
-    test_orbital_partitioner(universe);
 
     madness::finalize();
     return 0;
